fix: Restore lost idiom and quiz data
Restores the `idioms.json` and `quiz_questions.json` files to a version from a lost commit.

This recovers a significant amount of data that was accidentally lost due to a bad merge. The correct version was found by searching through dangling commits in the repository's history.
